% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rd.R
\name{is_Rd_newline}
\alias{is_Rd_newline}
\title{Check if an element is a newline}
\usage{
is_Rd_newline(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if an element is a newline
}
