% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{pieces}
\alias{pieces}
\alias{Rd.newline}
\alias{Rd.code.newline}
\alias{Rd.break}
\title{Rd Pieces}
\usage{
Rd.newline
Rd.code.newline
Rd.break
}
\description{
These objects are provided to help construct Rd documents.
}
