% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-package.R
\docType{package}
\name{rd-package}
\alias{rd}
\alias{rd-package}
\title{rd: Region Dalarnas Analyspaket (paraply)}
\description{
Paraplypaket för Region Dalarnas analysarbete. `library(rd)` drar in alla rd-analyspaket på en gång (rdverktyg, rddiagram, rdgis, rdpostgres, rdshinyappar). Deploy- och serverdriftspaketen (rddeploy, rdadminportal) ingår inte - de laddas explicit vid behov.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
