% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_extensions.R
\name{Rd_c}
\alias{Rd_c}
\title{Combine Rd elements}
\usage{
Rd_c(...)
}
\arguments{
\item{...}{\link{Rd} elements.}
}
\description{
Combine Rd elements
}
\note{
The special considerations for Rd elements necessitate a
special method for combining the traditional generic \link[base:c]{c}
cannot be used due to problems that occur, particularly in the
\href{http://rstudio.com}{RStudio} GUI.
}
