% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-Rd.R
\name{Rd_compact}
\alias{Rd_compact}
\title{Compact a list into an Rd vector}
\usage{
Rd_compact(l)
}
\arguments{
\item{l}{list of Rd objects.}
}
\description{
When creating Rd from other objects it is common to
iterate over a vector or list and convert each \href{toRd()}{to Rd}
then combine the results. However \code{\link[=toRd]{toRd()}} can return an \link{Rd_string},
\link{Rd_tag}, or an \link{Rd} container. to combine these together
intelligently, use \code{Rd_compact} which converts each to an Rd
container, placing strings and tags in a container, then
concatenating all together.
}
