% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd_lines.R
\name{Rd_lines}
\alias{Rd_lines}
\title{Combine with line breaks}
\usage{
Rd_lines(l, ...)
}
\arguments{
\item{l}{A list of valid Rd objects.}

\item{...}{Arguments passed on to \code{Rd_canonize}
\describe{
  \item{rd}{the Rd container object to put in canonical form.}
  \item{.check}{Perform input checks?}
}}
}
\description{
Combine a list of Rd objects with a line break between each.
}
