% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{Rd_split}
\alias{Rd_split}
\title{Split a rd object into relevant lines.}
\usage{
Rd_split(rd)
}
\arguments{
\item{rd}{an rd list.}
}
\description{
Split a rd object into relevant lines.
}
